user_service/pyproject.toml
[tool.poetry]
name = "project-x-591"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
sqlmodel = "^0.0.22"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
psycopg = "^3.2.3"
psycopg2 = "^2.9.10"
psycopg-binary = "^3.2.3"
python-multipart = "^0.0.17"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
